Avoid IP assignment duplication
Do not assign an IP to client/server interfaces, as it is expected to be
assigned already, in fact the 'get_server_client_ips_and_ifs' function
will fail the run if it is not assigned. Currently the same IP with
possibly a different subnet is assigned to the interfaces, which is not
good.

Signed-off-by: Leonid Bloch <lbloch@nvidia.com>