tcp: Use the first net device associated with an IB device
In a case when there are several network devices associated
with a single IB device (e.g. on a DPU, where there are PF,
PF representor, and one or more SFs related to a single IB
device) use the first one. Currently, on BF2/3, the first
one is the generic SF (not any of the specially created SFs,
e.g. created for VirtIO-net emulation).

In the future, if a finer tuning will be needed, a more
clever mechanism can be implemented for the net device
selection.

Signed-off-by: Leonid Bloch <lbloch@nvidia.com>