TCP: Verify that hosts don't have unpopulated NUMA nodes
Currently, runs fail on hosts where some NUMA nodes are unpopulated.
However a scenario where some nodes are actually unpopulated is quite
rare. Instead it's very likely that some nodes seem to be unpopulated
because of BIOS misconfiguration. So for now, just fail the run with
an informative message in cases of such a misconfiguration, instead
of letting it continue and fail with a bunch of cryptic messages.

Signed-off-by: Leonid Bloch <lbloch@nvidia.com>