rdma: Skip ib_send_lat with CUDA only if the message size is <32 bytes
ib_send_lat can actually work with CUDA, but with message sizes only
higher than 32 (otherwise libmlx5 will try to inline the data to a WQE,
which will cause the CPU to try to memcpy the CUDA buffer). So if before
all the ib_send_lat tests were skipped with CUDA, now only the ones with
the small message sizes will be.

Signed-off-by: Leonid Bloch <lbloch@nvidia.com>